auth_service/requirements.txt
fastapi==0.68.1
sqlalchemy==1.4.23
passlib[bcrypt]==1.7.4
pyjwt==2.3.0
aiohttp==3.7.4.post0
uvicorn==0.15.0
psycopg2-binary==2.9.1
kafka-python==2.0.2
python-jose==3.3.0
python-multipart==0.0.5